Feature fin 1394

See merge request cfs/fnac!460

+++ service/src/main/java/ir/demisco/cfs/service/impl/DefaultAccountRelatedDescription.java
@@ -30,7 +30,10 @@ public class DefaultAccountRelatedDescription implements AccountRelatedDescripti
     @Override
     @Transactional(rollbackOn = Throwable.class)
     public boolean deleteAccountRelatedDescriptionById(Long accountRelatedDescriptionId) {
-        accountRelatedDescriptionRepository.findById(accountRelatedDescriptionId).orElseThrow(() -> new RuleException("fin.accountRelatedDescription.delete"));
+        if (!accountRelatedDescriptionRepository.findById(accountRelatedDescriptionId).isPresent()) {
+            throw new RuleException("fin.accountRelatedDescription.delete");
+        }
+
         accountRelatedDescriptionRepository.deleteById(accountRelatedDescriptionId);
         return true;
     }
@@ -66,9 +69,8 @@ public class DefaultAccountRelatedDescription implements AccountRelatedDescripti
                 return convertAccountRelatedDescriptionDto(accountRelatedDescription);
             }
         } else {
-
+            throw new RuleException("fin.financialAccount.notInformation");
         }
-        return null;
     }
